refac: HttpResponse 패키지 이동 및 빌더 패턴 적용
